+++ R/github.R
@@ -474,10 +474,20 @@ git_remote_connect <- function(repo, remote_repo) {
   with_cli_try("Connecting to remote repository {.val {remote_repo}}", {
     if (is.null(remote_repo))
       stop()
-
-    ownr <- gh::gh_whoami()$login
-    repo_name <- paste0(ownr, "/", remote_repo)
-    repo_url <- paste0("https://github.com/", repo_name)
+    # Check if remote_repo is a name or a URL
+    if(grepl("^http.+?/.+/.{0,}", remote_repo)){
+      own_repo <- gsub("^.+/(.+?/.+?)(\\.git)?$", "\\1", remote_repo)
+      if(length(gregexpr("/", own_repo)[[1]]) > 1) stop("Could not parse Git remote URL.")
+      ownr <- strsplit(own_repo, "/", fixed = TRUE)[[1]]
+      repo_name <- ownr[2]
+      ownr <- ownr[1]
+      repo_url <- remote_repo
+    } else {
+      ownr <- gh::gh_whoami()$login
+      repo_name <- paste0(ownr, "/", remote_repo)
+      repo_url <- paste0("https://github.com/", repo_name)
+    }
+    # Tothier
     test_repo <- try(gert::git_remote_ls(remote = repo_url), silent = TRUE)
     repo_exists <- isFALSE(inherits(test_repo, "try-error"))
     if (!repo_exists)