+++ R/load_entrypoint.R
@@ -26,7 +26,7 @@
 load_entrypoint <- function(worcs_directory = ".", verbose = TRUE, ...){
   cl <- as.list(match.call()[-1])
   # Filenames housekeeping
-  dn_worcs <- dirname(check_recursive(file.path(normalizePath(worcs_directory), ".worcs")))
+  dn_worcs <- worcs_root(path = worcs_directory)
   fn_worcs <- file.path(dn_worcs, ".worcs")
   if(file.exists(fn_worcs)){
     worcsfile <- read_yaml(fn_worcs)